LPU_Data_Analytic_Tool/requirements.txt
streamlit
pandas
gspread
google-auth
google-api-python-client
isodate
apify-client
